requirements.txt
firebase_admin==7.3.0
Flask==3.1.3
flask-smorest==0.47.0
python-dotenv==1.2.2
Flask-Limiter==4.1.1
redis==7.4.0